+++ tests/testthat/test-write_sumstats.R
@@ -4,31 +4,28 @@ test_that("write_sumstats works", {
     is_32bit_windows <- .Platform$OS.type == "windows" ##&&
         ##.Platform$r_arch == "i386"
     if (!is_32bit_windows) {
-        sumstats_dt <- MungeSumstats:::formatted_example()
+        sumstats_dt <- MungeSumstats::formatted_example()
         
         run_tests <- function(sumstats_dt,
                               fileext,
                               tabix_index=FALSE,
                               write_vcf=FALSE,
-                              standardise_headers=FALSE){
+                              standardise_headers=FALSE,
+                              save_path_check=FALSE){
             message(fileext) 
-            #### Check paths ####
-            path_in <- tempfile(fileext = fileext)
-            check <- MungeSumstats:::check_save_path(save_path = path_in, 
-                                                     log_folder = tempdir(),
-                                                     log_folder_ind = FALSE,
-                                                     tabix_index = tabix_index, 
-                                                     write_vcf = write_vcf)
-            path_in <- check$save_path
-            
+            #### Check paths #### 
+            message("=== write tests ===")
             path_out <- MungeSumstats::write_sumstats(
                 sumstats_dt = sumstats_dt,
-                save_path = path_in,
+                save_path = tempfile(fileext = fileext),
                 write_vcf = write_vcf,
                 tabix_index = tabix_index,
-                return_path = TRUE
+                return_path = TRUE, 
+                save_path_check = save_path_check
             )
             testthat::expect_true(file.exists(path_out))
+            
+            message("\n=== read tests ===")
             dat <- MungeSumstats::read_sumstats(
                 path = path_out, 
                 standardise_headers = standardise_headers)
@@ -40,7 +37,9 @@ test_that("write_sumstats works", {
         dat <- run_tests(sumstats_dt = sumstats_dt, 
                          fileext = ".tsv")
         dat <- run_tests(sumstats_dt = sumstats_dt, 
-                         fileext = ".tsv.gz")
+                         fileext = ".tsv.gz") 
+        dat <- run_tests(sumstats_dt = sumstats_dt, 
+                         fileext = ".tsv.bgz")
         dat <- run_tests(sumstats_dt = sumstats_dt, 
                          fileext = ".tsv.gz", 
                          tabix_index = TRUE)
@@ -51,22 +50,35 @@ test_that("write_sumstats works", {
                          fileext = ".csv")
         dat <- run_tests(sumstats_dt = sumstats_dt, 
                          fileext = ".csv.gz")
-        
-        #### VCF formats ####
         # write_vcf=F
+        ### Will write as tsv even though specified vcf suffix
         dat <- run_tests(sumstats_dt = sumstats_dt, 
                          fileext = ".vcf") 
         dat <- run_tests(sumstats_dt = sumstats_dt, 
                          fileext = ".vcf.gz") 
+        
+        #### VCF formats ####
         # write_vcf=T
         dat <- run_tests(sumstats_dt = sumstats_dt, 
                          write_vcf = TRUE,
                          fileext = ".vcf") 
-        ## with indexing
         dat <- run_tests(sumstats_dt = sumstats_dt, 
                          write_vcf = TRUE,
-                         tabix_index = TRUE,
-                         fileext = ".vcf") 
+                         fileext = ".vcf.gz") 
+        ## with indexing
+        # providing the correct suffix is important
+       testthat::expect_failure(
+           dat <- run_tests(sumstats_dt = sumstats_dt, 
+                            write_vcf = TRUE,
+                            tabix_index = TRUE,
+                            fileext = ".vcf", 
+                            save_path_check = FALSE) 
+       )
+       dat <- run_tests(sumstats_dt = sumstats_dt, 
+                        write_vcf = TRUE,
+                        tabix_index = TRUE,
+                        fileext = ".vcf", 
+                        save_path_check = TRUE) 
         dat <- run_tests(sumstats_dt = sumstats_dt, 
                          tabix_index = TRUE,
                          write_vcf = TRUE,